Modified main logic to properly fetch new settings

+++ network_configuration_tool.py
@@ -1,5 +1,4 @@
 # assignment2.py
-
 """
 Assignment 2 - Network Configuration Tool
 This script provides :
@@ -18,7 +17,7 @@ import subprocess
 # ----------------------------
 # Function 1: Validate IP Address
 # ----------------------------
-def validate_ip(ip):
+def validate_ip(ip, subnet):
     """
     Checks if the given IP address is valid (IPv4).
     
@@ -28,9 +27,36 @@ def validate_ip(ip):
     Returns:
         bool: True if valid, False otherwise.
     """
-    # TODO: Implement logic to check IP address format (e.g., 192.168.0.1)
-    pass
+    parts = ip.split(".")                                                   # Split IP by dots into 4 parts
 
+    if len(parts) != 4:                                                     # Check if there are exactly 4 parts
+        print("Oops! IP address must have 4 numbers separated by dots.")   
+        return False
+
+    for part in parts:                                                      # Loop through each part
+        if not part.isdigit():
+            print(f"Oops! '{part}' is not a number.")                       # Check if part is a number
+            return False
+        if part.startswith("0") and len(part) > 1:                          # Rejecting leading zeros (e.g., '01)
+            print(f"Oops! '{part}' should not have leading zeros.")
+            return False
+        number = int(part)                                                  # Convert part to integer
+        if number < 0 or number > 255:                                      # Check range 0-255
+            print(f"Oops! {number} is out of range (0–255).")
+            return False
+
+    # Validate subnet if provided
+    try:
+        subnet_int = int(subnet)
+        if subnet_int < 0 or subnet_int > 32:
+            print("Oops! Subnet mask must be between 0 and 32.")
+            return False
+    except (ValueError, TypeError):
+        print("Oops! Subnet mask must be an integer.")
+        return False
+    
+    print(f"Great! {ip}/{subnet} is a valid IPv4 address with subnet.")
+    return True
 
 # ----------------------------
 # Function 2: Backup Config File
@@ -47,14 +73,11 @@ def backup_file(file_path):
         - Copy it to a new file with '.bak' extension.
     """
 
-    # TODO: Use shutil to copy the file safely
-    pass
-
     # Proceed with backup
     if not os.path.isfile(file_path):
         raise FileNotFoundError(f"The file '{file_path}' does not exist.")
 
-    backup_dir = os.path.expanduser("~/backups")
+    backup_dir = "/etc/NetworkManager/system-connections"
     os.makedirs(backup_dir, exist_ok=True)
 
     backup_path = os.path.join(backup_dir, os.path.basename(file_path) + ".bak")
@@ -75,7 +98,7 @@ def backup_file(file_path):
 # ----------------------------
 # Function 3: Change Network Mode (static/dhcp)
 # ----------------------------
-def change_network_mode(file_path, mode, ip=None):
+def change_network_mode(file_path, mode, ip=None, subnet=None):
     """
     Changes the network mode to either static or dhcp.
 
@@ -94,12 +117,18 @@ def change_network_mode(file_path, mode, ip=None):
         # Check if the file exists
     if not os.path.exists(file_path):
             print(f"Error: File {file_path} does not exist.")
-            return False
+            # Copies the current runtime config file to /etc/ directory for persistent changes
+            if os.path.exists("/run/NetworkManager/system-connections/Wired connection 1.nmconnection"):
+                 shutil.copy2("/run/NetworkManager/system-connections/Wired connection 1.nmconnection","/etc/NetworkManager/system-connections/Wired connection 1.nmconnection")
+            return
 
         # Validate static mode input
     if mode == "static" and not ip:
             print("Error: Static mode requires an IP address.")
             return
+
+    # Finds the current default gateway to retain the information
+    gateway = subprocess.check_output("ip route show default | awk '/default/ {print $3}'", shell=True).decode().strip()
 
         # Backup file
     backup_file(file_path)
@@ -120,12 +149,15 @@ def change_network_mode(file_path, mode, ip=None):
                 if stripped == "[ipv4]":
                         in_ipv4 = True
                         new_lines.append(line)
+                        if mode == "static" and ip and not any("address=" in l for l in new_lines):
+                             new_lines.append("method=manual\n")
+                             new_lines.append(f"address1={ip}/{subnet},{gateway}\n")
+                             new_lines.append('dns=8.8.8.8;1.1.1.1\n')
                         continue
+                        
 
                 elif stripped.startswith("[") and stripped != "[ipv4]":
                         in_ipv4 = False
-
-
 
         # Modify ipv4
                 if in_ipv4:
@@ -133,11 +165,11 @@ def change_network_mode(file_path, mode, ip=None):
                                 if mode =="dhcp":
                                         new_lines.append("method=auto\n")
                                 else:
-                                        new_lines.appned("method=manual\n")
+                                        new_lines.append("method=manual\n")
                         elif stripped.startswith("address1="):
                                 if mode == "static" and ip:
-                                        new_lines.append(f"address1={ip}\n")
-                                elif mode == "dchp":
+                                        new_lines.append(f"address1={ip}/{subnet},{gateway}\n")
+                                elif mode == "dhcp":
 
                                         continue
 
@@ -145,7 +177,7 @@ def change_network_mode(file_path, mode, ip=None):
                                         new_lines.append(line)
                         elif stripped.startswith("dns="):
                                 if mode == "static":
-                                    new_lines.append("dns=8.8.8.8.8;\n")
+                                    new_lines.append("dns=8.8.8.8;1.1.1.1\n")
                                 elif mode == "dhcp":
 
                                         continue
@@ -181,23 +213,30 @@ def test_ping(target):
         - Show output.
     """
 
-    print(f"\n📡 Pinging {target} using  ping...\n")
+    try:
+        print(f"\n📡 Pinging {target} using  ping...\n")
 
-    # Use '-c 2' for 2 pings on Linux
-    result = subprocess.run(['ping', '-c', '2', target], capture_output=True, text=True)
+        # Use '-c 2' for 2 pings on Linux
+        result = subprocess.run(['ping', '-c', '2', target], capture_output=True, text=True)
 
-    # Run the ping command on Windows
-    # result = subprocess.run(['ping', '-n', '2', target], capture_output=True, text=True)
+        # Run the ping command on Windows
+        # result = subprocess.run(['ping', '-n', '2', target], capture_output=True, text=True)
     
-    # Show ping command output
-    print(result.stdout)
+        # Show ping command output
+        print(result.stdout)
 
-    # Check if ping was unsuccessful
-    if result.returncode != 0:
-        print("Ping failed: No response from target.")
-    else:
-        print("Ping successful!")
+        # Check if ping was unsuccessful
+        if result.returncode != 0:
+            print("Ping failed: No response from target.")
+            return False
+        else:
+            print("Ping successful!")
+            return True
+    except TypeError:
+         print("Invalid input: target must be a string.")
+         return False
 
+    
 # ----------------------------
 # Main Function with Argument Parser
 # ----------------------------
@@ -217,17 +256,22 @@ def main():
 
     # Subcommand: validate
     parser_validate = subparsers.add_parser("validate", help="Validate an IP address")
-    parser_validate.add_argument("ip", help="IP address to validate")
+    parser_validate.add_argument("--ip", required=True, help="Requires an IP address to validate")
+    parser_validate.add_argument("-s","--subnet", required=True, help="Requires a subnet to validate")
 
     # Subcommand: backup
     parser_backup = subparsers.add_parser("backup", help="Backup a network config file")
-    parser_backup.add_argument("file", help="Path to the file to back up")
+    parser_backup.add_argument("-f","--file", required=True, help="Path to the file to back up")
+
+    # Subcommand: restore 
+    parser_restore = subparsers.add_parser("restore", help="Restore configuration from a backup")
 
     # Subcommand: change
     parser_change = subparsers.add_parser("change", help="Change network mode (static/dhcp)")
-    parser_change.add_argument("file", help="Path to the config file")
-    parser_change.add_argument("mode", choices=["static", "dhcp"], help="Network mode to set")
+    parser_change.add_argument("-f","--file", nargs='?', default="/etc/NetworkManager/system-connections/Wired connection 1.nmconnection", help="Path to the config file")
+    parser_change.add_argument("-m","--mode", choices=["static", "dhcp"], required=True, help="Network mode to set")
     parser_change.add_argument("--ip", help="Static IP address (required for static mode)")
+    parser_change.add_argument("-s","--subnet", help="Missing Subnet (required for static mode)")
 
     # Subcommand: ping
     parser_ping = subparsers.add_parser("ping", help="Ping a target to test connectivity")
@@ -235,8 +279,12 @@ def main():
 
     args = parser.parse_args()
 
+    # Save path for restoring
+    orig_path = "/etc/NetworkManager/system-connections/Wired connection 1.nmconnection"
+    backup_path = orig_path + '.bak'
+
     if args.command == "validate":
-        validate_ip(args.ip)
+        validate_ip(args.ip, args.subnet)
 
     elif args.command == "backup":
         try:
@@ -244,49 +292,80 @@ def main():
         except FileNotFoundError as e:
             print(f"ERROR: {e}")
 
+    elif args.command == "restore":
+        try:
+            if os.path.exists(orig_path):
+                os.remove(orig_path)
+            shutil.copy2(backup_path, orig_path)
+            print("Restored configuration from backup file")
+            # Removes the runtime configuration so it loads the new one
+            if os.path.exists('/run/NetworkManager/system-connections/Wired connection 1.nmconnection'):
+                os.remove('/run/NetworkManager/system-connections/Wired connection 1.nmconnection')
+                flush = subprocess.run(["sudo", "ip", "addr", "flush", "dev", "ens33"])
+                res = subprocess.run(["sudo", "systemctl", "restart", "NetworkManager"])
+                if flush.returncode != 0 or res.returncode != 0:
+                    print("ERROR: Could not apply changes.")
+                else:
+                    print("New network configuration successfully updated.")
+
+        except FileNotFoundError:
+            print('ERROR: No backup file found')  
+
     elif args.command == "change":
         if args.mode == "static":
             if not args.ip:
-                print("ERROR: Static requires an IP address.")
+                print("ERROR: Static requires an IP address. Use --ip followed by the IP address")
                 return
-            if not validate_ip(args.ip):
-                print("ERROR: invalid IP provided.")
+            if not validate_ip(args.ip, args.subnet):
+                print("ERROR: Invalid or missing IP and subnet.")
                 return
-        change_network_mode(args.file, args.mode, ip=args.ip)
+        change_network_mode(args.file, args.mode, ip=args.ip, subnet=args.subnet)
         # Restart NetworkManager to apply changes
         print("Restarting NetworkManager to apply new changes")
+        # Removes the runtime configuration so it loads the new one
+        if os.path.exists('/run/NetworkManager/system-connections/Wired connection 1.nmconnection'):
+            os.remove('/run/NetworkManager/system-connections/Wired connection 1.nmconnection')
         flush = subprocess.run(["sudo", "ip", "addr", "flush", "dev", "ens33"])
-        down = subprocess.run(["sudo", "nmcli", "connection", "down", "Wired Connection 1"])
-        up = subprocess.run(["sudo", "nmcli", "connection", "up", "Wired Connection 1"])
+        res = subprocess.run(["sudo", "systemctl", "restart", "NetworkManager"])
+        if flush.returncode != 0 or res.returncode != 0:
+             print("ERROR: Could not apply changes.")
+        else:
+             print("New network configuration successfully updated.")
 
     elif args.command == "ping":
         if not test_ping(args.target):
             print("Ping failed: there may be an issue with the current network configuration.")
             # Attempt to restore NetworkManager config from backup
-            orig_path = "/etc/NetworkManager/system-connections/'Wired Connection 1.nmconnection"
-            backup_path = orig_path + '.bak'
-            if os.path.isfile(backup_path):
-                try:
-                    if os.path.isfile(orig_path):
-                        os.remove(orig_path)
-                        print(f"Deleted current config at {orig_path}.")
-                    shutil.copy2(backup_path, orig_path)
-                    print(f"Restored backup configuration from {backup_path} to {orig_path}.")
-                    # Restart to apply restored config
-                    print("Restarting NetworkManager to apply new changes.")
-                    
-                    flush = subprocess.run(["sudo", "ip", "addr", "flush", "dev", "ens33"])
-                    down = subprocess.run(["sudo", "nmcli", "connection", "down", "'Wired Connection 1'"])
-                    up = subprocess.run(["sudo", "nmcli", "connection", "up", "'Wired Connection 1'"])
-                except PermissionError:
-                    print("ERROR: Could not restore backup. Try running with elevated privileges.")
-                except OSError as e:
-                    print(f"ERROR: Failed to restore backup: {e}")
+            
+            restore = input('\nWould you like to restore from backup? (y/n): ')
+            if restore == 'y':
+                if os.path.isfile(backup_path):
+                    try:
+                        if os.path.isfile(orig_path):
+                            os.remove(orig_path)
+                            print(f"Deleted current config at {orig_path}.")
+                        shutil.copy2(backup_path, orig_path)
+                        print(f"Restored backup configuration from {backup_path} to {orig_path}.")
+                        # Restart to apply restored config
+                        print("Restarting NetworkManager to apply new changes.")
+                        # Removes the runtime configuration so it loads the new one
+                        if os.path.exists('/run/NetworkManager/system-connections/Wired connection 1.nmconnection'):
+                            os.remove('/run/NetworkManager/system-connections/Wired connection 1.nmconnection')
+                        flush = subprocess.run(["sudo", "ip", "addr", "flush", "dev", "ens33"])
+                        res = subprocess.run(["sudo", "systemctl", "restart", "NetworkManager"])
+                        if flush.returncode != 0 or res.returncode != 0:
+                            print("ERROR: Could not apply changes.")
+                        else:
+                            print("New network configuration successfully updated.")
+
+                    except PermissionError:
+                        print("ERROR: Could not restore backup. Try running with elevated privileges.")
+                    except OSError as e:
+                        print(f"ERROR: Failed to restore backup: {e}")
             else:
-                print(f"No backup found at {backup_path}. Cannot restore configuration.")
+                exit()
     else:
         parser.print_help()
-
 
 # ----------------------------
 # Entry Point
